+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Add `ContainerCreateOptsBuilder::mounts`
 * Add `ContainerCreateOptsBuilder::passwd_entry`
 * Add `ImageBuildOptsBuilder::all_platforms`
+* Add `ImageBuildOptsBuilder::unset_env`
 
 # 0.7.0
 * Make futures returned by methods Send + Sync